refatoração de Código

+++ germina_app/lib/models/sensors/temp_sensor.dart
@@ -1,13 +1,13 @@
 import 'package:germina_app/models/sensors/sensor.dart';
 
 class TempSensor extends Sensor {
-  int temperature;
-  int umidity;
+  double temperature;
+  double umidity;
 
   TempSensor(this.temperature, this.umidity,
       {required String name,
       required String protocol,
       required String uri,
       String category = 'tempSensor'})
-      : super(name: name,protocol: protocol,uri: uri,category: category);
+      : super(name: name, protocol: protocol, uri: uri, category: category);
 }
